feat: context-aware channel rate limit with float RPM support
- Context-aware skip (other channels available) or 429 (all rate-limited)
- RateLimitRPM/DefaultRPM changed from int to float64, window auto-scales
- controller/relay.go: skip rate-limited channels via ExcludeChannel
- middleware/distributor.go: remove reject, defer to controller
- constant/context_key.go: add rate limit tracking keys
- relaykit/types/error.go: add ErrorCodeChannelRateLimited
- model/channel_cache.go: add GetChannelIDsForGroupModel

+++ relaykit/dto/channel_settings.go
@@ -91,7 +91,7 @@ type ChannelOtherSettings struct {
 	// "", "allow", "safe", "strict".
 	ToolLossPolicy string `json:"tool_loss_policy,omitempty"`
 	RateLimitEnabled                      bool                  `json:"rate_limit_enabled,omitempty"`
-	RateLimitRPM                          int                   `json:"rate_limit_rpm,omitempty"`
+	RateLimitRPM                          float64               `json:"rate_limit_rpm,omitempty"`
 	RateLimitTPM                          int                   `json:"rate_limit_tpm,omitempty"`
 }
